# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Meiro Decision & Eligibility Engine API
  version: 0.1.0
  description: REST API for real-time profile decision and eligibility evaluation.
servers:
  - url: http://localhost:8090
    description: Docker Compose local service
  - url: http://localhost:8080
    description: Direct Node.js local service
security:
  - bearerAuth: []
paths:
...
  /v1/evaluate:
    post:
      summary: Evaluate one profile against a published rule set
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
            examples:
              nextBestOffer:
                value:
                  decision_key: next_best_offer
                  profile_key: profile-123
                  identifiers:
                    - typeId: email
                      value: user@example.com
                  attributes:
                    lead_score:
                      - value: 82
                    web_engagement_score:
                      - value: 71
                    interacted_promotions:
                      - value: [retention_discount]
                    customer_lifetime_value:
                      - value: 12400
                    monetary_rfm:
                      - value: 5
                    churn_risk_score:
                      - value: 0.2
                    outstanding_balance_tier:
                      - value: low
                    late_payments_count_12m:
                      - value: 0
                    survey_nps_latest:
                      - value: 8
                    sustainability_score:
                      - value: 91
                  context:
                    channel: email
                    request_source: meiro_pipes
      responses:
        "200":
          description: Decision result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/evaluate/batch:
    post:
      summary: Evaluate up to 500 profiles
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision_key, profiles]
              properties:
                decision_key:
                  type: string
                rule_version:
                  type: integer
                  nullable: true
                profiles:
                  type: array
                  maxItems: 500
                  items:
                    $ref: "#/components/schemas/EvaluateRequestProfile"
      responses:
        "200":
          description: Batch decision results
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/EvaluateResponse"
  /v1/rule-sets:
    get:
      summary: List rule sets
      responses:
        "200":
          description: Rule set list
          content:
            application/json:
              schema:
                type: object
                required: [rule_sets]
                properties:
                  rule_sets:
                    type: array
                    items:
                      $ref: "#/components/schemas/RuleSetSummary"
    post:
      summary: Create a rule set draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleSetWrite"
      responses:
        "201":
          description: Created rule set
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule_set:
                    $ref: "#/components/schemas/RuleSetSummary"
...
  /v1/audit:
    get:
      summary: Query the evaluation audit log
      parameters:
        - name: decision_key
          in: query
          schema:
            type: string
        - name: profile_key
          in: query
          schema:
            type: string
        - name: result
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit:
                    type: array
                    items:
                      type: object
            text/csv:
              schema:
                type: string
...
  /v1/schema:
    get:
      summary: List cached Meiro schema items
      parameters:
        - name: kind
          in: query
          schema:
            type: string
            enum: [attribute, segment, context]
      responses:
        "200":
          description: Cached schema items
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: array
                    items:
                      $ref: "#/components/schemas/SchemaItem"
  /v1/schema/import:
    post:
      summary: Manually import Meiro schema items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchemaImport"
      responses:
        "200":
          description: Imported schema items by kind
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: object
  /v1/schema/sync:
    post:
      summary: Sync schema by sampling a Meiro profile
      description: Calls the configured Meiro Profile API endpoint with an identifier and infers attribute, segment, and context schema from the returned profile payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier_type, identifier_value]
              properties:
                identifier_type:
                  type: string
                identifier_value:
                  type: string
                meiro_api_url:
                  type: string
                meiro_api_token:
                  type: string
      responses:
        "200":
          description: Imported schema items inferred from a profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: string
                  imported:
                    type: object
                  profile_shape:
                    type: array
                    items:
                      type: string
  /v1/export:
    get:
      summary: Export rule sets and lookup tables as a promotion bundle
      parameters:
        - name: include_audit
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Configuration bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigBundle"
  /v1/import:
    post:
      summary: Import a promotion bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigBundle"
      responses:
        "200":
          description: Import summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: object
                    properties:
                      rule_sets:
                        type: integer
                      lookup_tables:
                        type: integer
  /v1/tokens:
    get:
      summary: List API tokens
      responses:
        "200":
          description: API tokens without plaintext secret values
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiToken"
    post:
      summary: Create an API token
      description: The plaintext token is returned once and is never shown again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [viewer, editor, publisher, admin, evaluate]
      responses:
        "201":
          description: Created token with one-time plaintext value
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    allOf:
                      - $ref: "#/components/schemas/ApiToken"
                      - type: object
                        properties:
                          token:
                            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
    Identifier:
      type: object
      required: [typeId, value]
      properties:
        typeId:
          type: string
        value:
          type: string
...
    EvaluateRequest:
      allOf:
        - type: object
          required: [decision_key]
          properties:
            decision_key:
              type: string
            rule_version:
              type: integer
              nullable: true
        - $ref: "#/components/schemas/EvaluateRequestProfile"
    EvaluateRequestProfile:
      type: object
      required: [profile_key, identifiers]
      properties:
        profile_key:
          type: string
        identifiers:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Identifier"
        attributes:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/AttributeValue"
        segments:
          type: object
          additionalProperties:
            type: boolean
        context:
          type: object
          additionalProperties: true
    EvaluateResponse:
      type: object
      required: [decision_key, profile_key, evaluated_at, rule_version, result, outputs, matched_rules, errors]
      properties:
        decision_key:
          type: string
        profile_key:
          type: string
        evaluated_at:
          type: string
          format: date-time
        rule_version:
          type: integer
        result:
          type: string
          example: eligible
        outputs:
          type: object
          additionalProperties: true
        matched_rules:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
...